refactor: move components to `src`
