+++ app/[lang]/recent-list/GithubUserList.tsx
@@ -5,17 +5,16 @@ import {useMemo, useRef, useState} from 'react';
 import clsx from 'clsx';
 import Image from 'next/image';
 
-import type {UserListItem} from '../../../src/fetches/recentList';
-import {fetchRecentList} from '../../../src/fetches/recentList';
-import type {Translates} from '../../../src/localization';
-import type {ColumnDef} from '../(common)/DataTable';
-import {DataTable} from '../(common)/DataTable';
-import styles from '../styles.module.css';
-
 import type {Tier} from './TierRowItem';
 import TierRowItem from './TierRowItem';
 
-import {H4, H5} from '~/components/Typography';
+import type {ColumnDef} from '~/components/uis/DataTable';
+import {DataTable} from '~/components/uis/DataTable';
+import {H4, H5} from '~/components/uis/Typography';
+import type {UserListItem} from '~/src/apis/recentList';
+import {fetchRecentList} from '~/src/apis/recentList';
+import type {Translates} from '~/src/localization';
+import styles from '~/styles.module.css';
 
 type Props = {
   t: Translates['recentList'];
