+++ app/[lang]/recent-list/TierRowItem.tsx
@@ -1,7 +1,7 @@
 import type {ReactElement} from 'react';
 import Image from 'next/image';
 
-import {getTierSvg} from '../../../src/utils/functions';
+import {getTierSvg} from '~/src/utils/functions';
 
 export type Tier =
   | 'Iron'
