+++ app/[lang]/stats/[login]/Scouter/StatsDetails/SectionEarth.tsx
@@ -2,11 +2,10 @@ import type {ReactElement} from 'react';
 import clsx from 'clsx';
 import {Inter} from 'next/font/google';
 
-import type {StatsDetail} from '../../../../../../server/plugins';
-import type {DoobooStatsResponse} from '../../../../../../server/services/githubService';
-import type {Translates} from '../../../../../../src/localization';
-
 import SvgEnter from '@/public/assets/enter.svg';
+import type {StatsDetail} from '~/server/plugins';
+import type {DoobooStatsResponse} from '~/server/services/githubService';
+import type {Translates} from '~/src/localization';
 
 const inter = Inter({subsets: ['latin']});
