+++ app/[lang]/stats/[login]/Scouter/StatsDetails/SectionGold.tsx
@@ -2,9 +2,9 @@ import type {ReactElement} from 'react';
 import clsx from 'clsx';
 import {Inter} from 'next/font/google';
 
-import type {StatsDetail} from '../../../../../../server/plugins';
-import type {DoobooStatsResponse} from '../../../../../../server/services/githubService';
-import type {Translates} from '../../../../../../src/localization';
+import type {StatsDetail} from '~/server/plugins';
+import type {DoobooStatsResponse} from '~/server/services/githubService';
+import type {Translates} from '~/src/localization';
 
 const inter = Inter({subsets: ['latin']});
