+++ app/[lang]/stats/[login]/page.tsx
@@ -1,13 +1,13 @@
 import type {ReactElement} from 'react';
 
-import {getDoobooStats} from '../../../../server/services/githubService';
-import {getSupabaseServerComponentClient} from '../../../../server/supabaseClient';
-import {getTranslates} from '../../../../src/localization';
 import Container from '../Container';
 
 import Scouter from './Scouter';
 
 import type {Locale} from '~/i18n';
+import {getDoobooStats} from '~/server/services/githubService';
+import {getSupabaseServerComponentClient} from '~/server/supabaseClient';
+import {getTranslates} from '~/src/localization';
 
 type Props = {
   params: {lang: Locale; login: string};
