+++ src/apis/newsLetter.ts
