+++ tailwind.config.js
@@ -55,6 +55,7 @@ const shadowFill = plugin(
 module.exports = {
   content: [
     './app/**/*.{js,ts,jsx,tsx}',
+    './src/**/*.{js,ts,jsx,tsx}',
     './components/**/*.{js,ts,jsx,tsx}',
     '!./node_modules',
   ],
