+++ tsconfig.json
@@ -17,10 +17,14 @@
     "plugins": [{ "name": "next" }],
     "baseUrl": ".",
     "paths": {
-      "~/*": ["./src/*"],
-      "@/public/*": ["./public/*"],
-      "@/app/*": ["./app/[lang]/*"],
-      "@/server/*": ["./server/*"],
+      "~/*": ["./*"],
+      "~/public/*": ["./public/*"],
+      "~/assets/*": ["./assets/*"],
+      "~/styles/*": ["./styles/*"],
+      "~/app/*": ["./app/[lang]/*"],
+      "~/server/*": ["./server/*"],
+      "~/src/*": ["./src/*"],
+      "~/components/*": ["./src/components/*"],
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
